Make sure we can represent unsigned trilogy_binary_value_t's
This fixes the building and parsing code to check for and set a
`is_unsigned` boolean field on trilogy_binary_value_t. This lets
the caller specify the parameter value being passed should be
treated as unsigned on the server. As well as making sure we tell
the caller that a value received from the server is unsigned or
not.