ruby: use rb_interned_str when available
This API was introduced with Ruby 3.0
Ref: https://bugs.ruby-lang.org/issues/16029

It allows to re-used an existing interned string when possible.
Since column names are very often present in the codebase, the hitrate
is generally good.

One issue here however is that column names are created in ASCII because
the Ruby client doesn't know what the connection encoding is.

So it won't be able to re-use literal strings from the codebase, however
it will re-use the strings backing symbols (they are ASCII).

But even on misses, since `Hash#[]=` will always try to dedup the string
by always getting an interned string, we save that extra lookup later when using
`Result#each_hash`.