Copy the in_transaction flag into Trilogy::Result
This flag (SERVER_STATUS_IN_TRANS, position 0x0001) is part of the
server_status bitfield sent with every OK_Packet which is already stored
on the connection.  But this particular data is not always safe to read
from the connection since in some cases one or more queries may have
been executed since the result was generated.  So to increase safety
let's also store it on the Result itself.