Verify exact length of auth_data_len
Previously here we were checking for a maximum length, ensuring there
was no buffer overflow. However, we did not check for a minimum length.
In some cases, if the database server was to send the client a specially
crafted packet, this could result in reading from uninitialized memory.

This checks that the auth_data_len is exactly 21 and rejects other
values.