ruby: Use TypedData_* APIs
The `Data_*` API still works, and AFAIK are not officially deprecated
but they're certainly no longer recommended.

In the context of trilogy the difference in implementation is very marginal.

As it doesn't hold onto any `VALUE` reference, it can register itself as
write barrier protected, so it isn't marked on every minor GC runs.