Cast to Integer when there are no decimal digits
Prior to this commit we'd always cast a SUM query to `BigDecimal`, since
the column type we get from the server is 246 (i.e.
TRILOGY_TYPE_NEWDECIMAL).

After this commit we'll only cast to `BigDecimal` if there are digits in
the decimal place. If there are none (i.e. if column->decimals is 0), we
cast to `Integer` instead.

Closes https://github.com/github/trilogy/issues/36