Trilogy::ProtocolError should use SQL codes to determine error class
Instead of raising all ProtocolErrors as connection errors,
we use the SQL error code to determine the type of error to raise
when handling the Trilogy error in the Ruby bindings.

The mapping of error codes to error classes was taken from the mysql2
implementation: https://github.com/brianmario/mysql2/blob/master/lib/mysql2/error.rb#L13-L27

Note that we don't need to care about the 2xxx errors because those
are raised by the libmysql client, which Trilogy doesn't deal with.

Co-authored-by: Jean Boussier <jean.boussier@shopify.com>