Implement #abandon_results! for Ruby binding
A common pattern when using multi-statements is to quickly iterate through the
existing results sets from the server so as to put the client back into
a usable state. This is especially helpful when multi-statement is
being used for bulk inserts, where the result sets aren't particularly
relevant. This helper method simplifies that pattern.