Avoid protocol error on empty password auth switch
Prior to this commit test_connection_error, which creates a connection
with a username that doesn't exist and an empty password, was flaky on
mysql 8. Sometimes it passed with the expected BaseConnectionError, but
often it failed with a TRILOGY_PROTOCOL_VIOLATION.

The source of the failure was `read_auth_switch_packet`, which only
handles `mysql_native_password` and `caching_sha2_password`, whereas in
the case of this failure the attempted auth switch was to
`sha256_password`.

I'm not entirely sure why we get an auth switch packet sometimes and an
"access denied" error packet other times (I'm guessing it has something
to do with mysql 8 changing the default auth plugin to a caching plugin,
but that's really just a guess). But in the case of an empty password it
seems like we can handle the switch just fine. We've done a similar
password_length check in https://github.com/github/trilogy/pull/21.

With this commit test_connection_error passes consistently.