Have test_connection_error test user account with wrong password
This test was being skipped for reasons described in
https://github.com/github/trilogy/issues/43.
Instead of testing for a connection error by using a nonexistent user,
we can use an existing user with the wrong password. This appears to avoid
the authentication plugin flakiness, while still executing the code path we
care about related to an "Access denied" connection error.