Preserve DatabaseError constant
https://github.com/github/trilogy/pull/15 changed the error class for
TRILOGY_ERR from `DatabaseError` to `ProtocolError`. This commit
reintroduces `DatabaseError` as an ancestor of `ProtocolError` for
compatibility reasons, so that rescuing `DatabaseError` still works as
before.

Note that even with this change there are a number of error codes (see
ProtocolError::ERROR_CODES) that are no longer `DatabaseError`s, since
we have explicitly recategorized them as `BaseConnectionError`,
`TimeoutError`, or `QueryError`.