COM_SET_OPTION returns an EOF, parse it as an OK
See, https://github.com/mysql/mysql-server/blob/1bfe02bdad6604d54913c62614bde57a055c8332/sql/sql_parse.cc#L2377-L2397,
which clarifies that the packet type returned is an eof. We should use
the EOF identifer to discriminate the packet, but parse it as an OK
because that's the information it contains.