Add a query flag to toggle the type of casting done by SUM
In https://github.com/github/trilogy/pull/37 there was a change that updated the value of SUM to be Integer if what is
being summed are all Integers. It used to be that it will _always_ return a Decimal type.

When we pulled in that change, there were tests that failed because we were doing arthimitic operations on the result of
queries that have SUM. Ruby will have different results if you divide intergers and decimals.

In order to keep that behavior for us (GitHub) while allows other consumers to use SUM and expect the _right_ type we
are introducing a new query flag that we can set in our adapter. This query flag will be what will determine if we
should be casting the value of SUM to follow the _new_ behavior or roll back to the _old_ behavior