Raise all SystemCallErrors as Trilogy::Error.
This allows syscall errors to be rescued generically by Trilogy::Error,
or handled individually by consumers using their oriinal Errno classes.

To make this work, we subclass all of the SystemCallErrors and include
the base error module on each of them.

I've also removed the constructors from the existing Errno class subclasses
(Trilogy::TimeoutError, Trilogy::ConnectionClosedError, etc.). We never
set an error_code on syscall errors, so the constructors are not needed.