Ruby: Implement Client#discard to cleanly dispose of connection post fork
If you use `Client#close` on a connection that was inherited from a parent
it will close the connection for the parent.

Instead we need to use some POSIX magic to close the file descriptor
without emitting any close notification, so that the server isn't mislead
into closing the connection on its side.