Allow disabling multi_results capability
https://github.com/github/trilogy/pull/57 set multi-results capability
as a default.

To safely upgrade to the latest version of Trilogy, we want to turn that
option off and then turn it on one cluster at a time.

This commit removes the multi-results capability default from
TRILOGY_CAPABILITIES_CLIENT but will still set it unless the client is
initialized with `multi_results: false`.

We had an option to enable
multi results, but removed it in
https://github.com/github/trilogy/pull/68 because the option was always
a no-op. This PR adds the option back, with some modification.

Co-authored-by: Daniel Colson <composerinteralia@github.com>