Split Ruby code into multiple files
This makes things a little easier to read, and also makes things a bit
more flexible. The extension needs the errors and `Result` to be defined
ahead of time, but the encoding stuff needs the charset constants that
are defined in the extension. We were getting around than before by setting
`@mysql_encodings_map` on the connection, but now we can require the
files in order so the constants are available as needed.