Move errno detection into _cb_raw_{write,read}
Previously we would inspect errno inside of client.c to try to return
TRILOGY_AGAIN in cases it seemed appropriate. However on an SSL socket
(or any other hypothetical generic socket) we really should not retry
unless OpenSSL tells us to.

This moves the logic into _cb_raw_*, so that it only applies to raw
sockets.