Introduce Trilogy::EOFError
This is a new Ruby error to represent the TRILOGY_CLOSED_CONNECTION
status code, which is the status returned when read returns 0, write
hits an EPIPE, or when attempting to communicate through a socket after
it was shutdown/closed due to a previous error.

This subclasses BaseConnectionError so that it can be caught by `rescue
Trilogy::ConnectionError` (which is what we'd like to recommend for
detecting network/communication issues) and will always include the
TRILOGY_CLOSED_CONNECTION as part of the message for backwards
compatibility, as that's how currently Rails and others have been
detecting it.

Previously this would have been raised as a QueryError.