Add field_type information to result
Towards #99, this exposes additional field data on the query result
via a `Result#columns` method. We were already using this data in the
C extension to determine how to cast the column values. We now add a pointer
to the `column_info` struct onto the result and then lazily add the values
to the `Result` instance when called.

This method could eventually replace the `#fields` method which at the moment only
includes the column name.

Co-authored-by: Daniel Colson <composerinteralia@github.com>
Co-authored-by: Charlotte Wen <ywenc@github.com>