Use symbols for type, charset
We want to set some of these values as symbols instead of ints
on the Column to make it easier to work with. In this commit, we
created loopups for int to string using macros, then changed the
type to a symbol via `rb_intern` + `rb_id2sym`. It's possible there's
a single function to do this, but we couldn't find one.

I also added a downcase function to make sure we get the values in
lower instead of uppercase. I'm sure there is a better way to do this.
I have no idea how to write C.

Co-authored-by: Daniel Colson <composerinteralia@github.com>