Hold struct reference in separate object
Previously, we had added a pointer to our column_info struct onto
our `Result` for lazy loading the data. This resulted in changing the
type of `Result` to `T_DATA` which could have negative performance
implications. Instead, we can create a separate `Columns` class to
store the pointer to our struct. A reference to `columns` now gets
eagerly loaded onto the result, but we only populate the actual information
for the columns when calling `Result#columns`. This seems like a good
trade off for lazy loading (vs. putting the pointer directly on `Result`).

Having said that, it's probably worth asking if the lazy loading is worth this
extra complexity at all, or if we shoudl simply populate `@columns` with an
already loaded array for the `Result` immediately.