Report CLOSED_CONNECTION for SSL_ERROR_ZERO_RETURN
If a server closes the SSL connection cleanly (SSL_shutdown) as we
attempt a read or write (ie. a clean SSL shutdown but dirty MySQL
shutdown) we will get SSL_ERROR_ZERO_RETURN as the return code.

Previously we would have reported this as:

    Trilogy::SSLError: trilogy_query_recv: SSL Error: (null)

Now it will be reported as:

    Trilogy::EOFError: trilogy_query_recv: TRILOGY_CLOSED_CONNECTION

On older versions (< 2.5.0) we may also have reported this as
TRILOGY_CLOSED_CONNECTION.

Co-authored-by: Brendan Dougherty <b.w.dougherty@gmail.com>