Call `getaddrinfo` from Ruby
`getaddrinfo` is notorious for not having a timeout, which can make
it hang for a very long time. Since Trilogy releases the GVL before
calling it, that can cause the whole VM to be unresponsive and no
longer respond to ctrl+c etc.

In 3.3.0, Ruby is now doing name resolution from a background thread,
making `Socket.getaddrinfo` interruptible. So by doing the name resolution
on the Ruby side, we should avoid such scenarios.