Use connect_timeout for initial connection
After the initial non-blocking connection call, we wait for the socket
to become writable as a signal that it has been successful (non-blocking
connects otherwise return immediately).

Previously this "writable" check used the write timeout instead of the
connect timeout, which may have been confusing for users. IMO this was a
bug introduced at some point prior to public release (at one time
the writable check didn't exist and the connect_timeout worked as one
would expect).

This commit adds a new TRILOGY_WAIT_CONNECT, which uses connect_timeout
to wait for the connection to become writable. If connect_timeout isn't
set it will use write_timeout for backwards compatibility.

Co-authored-by: Daniel Colson <composerinteralia@github.com>