Create and delete auth users in tests
In local dev, if you didn't manually create these users, the tests would
fail. This refactors the tests to add a new method that will create the
user with the right auth plugin and then delete it at the end.

I've also deleted the docker entrypoint sql scripts because now we're
creating the users.