Restore resetting sequence_number in connect_send
While it does not make sense to reset the sequence number in the middle
of parsing the handshake packet, it is reasonable to continue resetting
the sequence number once immediately when the connection is established.
As mentioned during review, this makes the send/recv of establishing the
connection similar to other send/recv pairs where the sequence number
gets reset during the send portion.