Add Rails setup to README
Previously, there was no documentation for how to use the adapter in Rails. With a local MySQL install it's pretty easy - you swap the `adapter` and Trilogy will use the `socket` just like `mysql2` does. But if the app is in Docker, a simple `adapter` change isn't enough; you'll get an `caching_sha2_password requires either TCP with TLS or a unix socket` error that does not tell you how to set up TCP with TLS. You need to set `ssl_mode`, too.

This commit adds a new section to the `contrib/ruby/README.md` to make all this clear.