Establish socket connection in Ruby
This moves the responsability for establishing the initial TCP or unix
socket connection to Ruby. This allows us to lean on Ruby's
implementation and configuration of DNS timeouts as well as happy
eyeballs/fast fallback.

In addition to improved timeouts, this should also deliver more detailed
and specific error messages.

Co-authored-by: Adam Hess <adamhess1991@gmail.com>
Co-authored-by: Jean Boussier <jean.boussier@gmail.com>