free_trilogy: handle the case where the buffer pool has been freed first
If the current Ractor is being GCed or the entire VM is, the buffer
pool instance may be garbage collected before the `Trilogy` instances.

There was code to handle it, but it wasn't correct, because
in such case `TypedData_Get_Struct` tries to raise an exception
which requires an allocation and that is illegal during garbage
collection.

So instead we need to manually check if the pool object is still
alive.