ruby: Prevent concurrent usage of Trilogy connections
Using a trology connection in a concurrent way will likely
lead to various protocol errors, but can also lead to various
crashes.

As such it is preferable if Trilogy raises a clear to understand
and deterministic error.

I chose to implement this on the Ruby side, because it's simpler and
more straightforward.

I initially tried to do it in C, but the `_cb_wait` callback may
raise or throw, so we'd need to use `rb_protect`, and release the lock
there. The problem is the callback doesn't have access to the
`trilogy_context` struct, so can't release the lock.

If we're adament this should be done in C, we'll need a much larger
refactor.