Test mysql 8.4
There are no debian tags for 8.4, so switch to non-debian tags which use
`oracle-linux:9-slim`. Unfortunately I couldn't figure out how to build
`mysql-community-test` on that so this commit is still pulling the test
plugins from 8.0-debian. That's sure to break eventually...

We'll keep testing with mysql_native_password for now on 8.4, which we
can do by having a different build.cnf for the two versions.