Don't bother checking in buffer pool on free
There's no guarantee that gc runs on the same Ractor which checked out
the pool, and I don't think it's helpful to give the buffers to an
arbitrary Ractor (which might not even be making any connections).

In the case we're allocating and freeing Trilogy connections anyways, I
we're going to be doing a malloc/free anyways, so the extra expense of
allocating the buffer shouldn't be too significant.